rename to ts & merge public apis
